Consolidate the mention feature into prompt-editor.tsx
Reviewability pass: prompt-mentions.ts and prompt-mention-context.ts fold
into the one file that owns the feature; the textarea-era API
(getActiveMention/applyMention) is dead since the Tiptap rewrite and is
deleted along with its tests and the redundant Mention renderText config.

Also fixes both open greptile findings on the multi-word resolution:
candidates whose last word lacks an extension dot no longer match unless
they consume the whole line (prose can't extend a mention into an
unintended longer filename), and the probe cap rises to 16 words with the
bound documented.
